Add script to get tx params; Add cronos network
