+++ .env.sample
@@ -1,1 +1,2 @@
 MNEMONIC=""
+RPC=""
