+++ artifacts/25/deployment.json
@@ -1,0 +1,7 @@
+{
+	"gasPrice": 10000000000000,
+	"gasLimit": 100000,
+	"signerAddress": "0xE1CB04A0fA36DdD16a06ea828007E35e1a3cBC37",
+	"transaction": "0xf8a6808609184e72a000830186a08080b853604580600e600039806000f350fe7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe03601600081602082378035828234f58015156039578182fd5b8082525050506014600cf355a054b6bc11ebcdd492b5083f82e745b27cf9aabd3afedc969a1fe863219c1094a9a00cd431f6d8e76c58b4ae488c96f95df8c57850088dd362a268cf051b226de259",
+	"address": "0x914d7Fec6aaC8cd542e72Bca78B30650d45643d7"
+}
