+++ artifacts/338/deployment.json
@@ -1,0 +1,7 @@
+{
+	"gasPrice": 10000000000000,
+	"gasLimit": 100000,
+	"signerAddress": "0xE1CB04A0fA36DdD16a06ea828007E35e1a3cBC37",
+	"transaction": "0xf8a8808609184e72a000830186a08080b853604580600e600039806000f350fe7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe03601600081602082378035828234f58015156039578182fd5b8082525050506014600cf38202c8a0749cead18bfc5151c8414fab334577f8c56120de5cda871c2237c855962184b3a042f0b6ed04459c15e37381167292f965a93f63e399f99b3c4f0f504b5b934dbb",
+	"address": "0x914d7Fec6aaC8cd542e72Bca78B30650d45643d7"
+}
