+++ scripts/compile.ts
@@ -4,6 +4,7 @@ import { CompilerOutput, CompilerInput, compileStandardWrapper, CompilerOutputCo
 import { ethers } from 'ethers'
 import dotenv from "dotenv";
 import yargs from 'yargs/yargs';
+import { compileContracts } from './utils';
 
 dotenv.config()
 
@@ -14,50 +15,6 @@ export async function ensureDirectoryExists(absoluteDirectoryPath: string) {
 		if (error.code === 'EEXIST') return
 		throw error
 	}
-}
-
-async function compileContracts(): Promise<CompilerOutput> {
-	const solidityFilePath = path.join(__dirname, '..', 'source', 'deterministic-deployment-proxy.yul')
-	const soliditySourceCode = await filesystem.readFile(solidityFilePath, 'utf8')
-	const compilerInput: CompilerInput = {
-		language: "Yul",
-		settings: {
-			optimizer: {
-				enabled: true,
-				details: {
-					yul: true,
-				},
-			},
-			outputSelection: {
-				"*": {
-					"*": [ "abi", "evm.bytecode.object", "evm.gasEstimates" ]
-				},
-			},
-		},
-		sources: {
-			'deterministic-deployment-proxy.yul': {
-				content: soliditySourceCode,
-			},
-		},
-	}
-	const compilerInputJson = JSON.stringify(compilerInput)
-	const compilerOutputJson = compileStandardWrapper(compilerInputJson)
-	const compilerOutput = JSON.parse(compilerOutputJson) as CompilerOutput
-	const errors = compilerOutput.errors
-	if (errors) {
-		let concatenatedErrors = "";
-
-		for (let error of errors) {
-			if (/Yul is still experimental/.test(error.message)) continue
-			concatenatedErrors += error.formattedMessage + "\n";
-		}
-
-		if (concatenatedErrors.length > 0) {
-			throw new Error("The following errors/warnings were returned by solc:\n\n" + concatenatedErrors);
-		}
-	}
-
-	return compilerOutput
 }
 
 async function writeBytecode(bytecode: string) {
