+++ scripts/utils.ts
@@ -1,0 +1,47 @@
+import { promises as filesystem } from 'fs'
+import * as path from 'path'
+import { CompilerOutput, CompilerInput, compileStandardWrapper } from 'solc'
+
+export async function compileContracts(): Promise<CompilerOutput> {
+	const solidityFilePath = path.join(__dirname, '..', 'source', 'deterministic-deployment-proxy.yul')
+	const soliditySourceCode = await filesystem.readFile(solidityFilePath, 'utf8')
+	const compilerInput: CompilerInput = {
+		language: "Yul",
+		settings: {
+			optimizer: {
+				enabled: true,
+				details: {
+					yul: true,
+				},
+			},
+			outputSelection: {
+				"*": {
+					"*": [ "abi", "evm.bytecode.object", "evm.gasEstimates" ]
+				},
+			},
+		},
+		sources: {
+			'deterministic-deployment-proxy.yul': {
+				content: soliditySourceCode,
+			},
+		},
+	}
+	const compilerInputJson = JSON.stringify(compilerInput)
+	const compilerOutputJson = compileStandardWrapper(compilerInputJson)
+	const compilerOutput = JSON.parse(compilerOutputJson) as CompilerOutput
+	const errors = compilerOutput.errors
+	if (errors) {
+		let concatenatedErrors = "";
+
+		for (let error of errors) {
+			if (/Yul is still experimental/.test(error.message)) continue
+			concatenatedErrors += error.formattedMessage + "\n";
+		}
+
+		if (concatenatedErrors.length > 0) {
+			throw new Error("The following errors/warnings were returned by solc:\n\n" + concatenatedErrors);
+		}
+	}
+
+	return compilerOutput
+}
